Move settings out and stop using compiled prefabs.
Compiled prefabs do not work when converted to .exe so bugs in prefab_classes may only cause issues when compiled.
