+++ src/splitguides/hotkeys/__init__.py
@@ -1,4 +1,1 @@
-from prefab_classes.compiled import prefab_compiler
-
-with prefab_compiler():
-    from .keyboard_fixer import Hotkey, read_hotkey, hotkey_or_none
+from .keyboard_fixer import Hotkey, read_hotkey, hotkey_or_none
