+++ src/splitguides/livesplit_client/messaging.py
@@ -8,7 +8,7 @@ from datetime import timedelta
 from prefab_classes import prefab, attribute
 
 
-@prefab(compile_prefab=True, compile_fallback=True)
+@prefab
 class LivesplitMessaging:
     connection = attribute()
