+++ src/splitguides/settings.py
@@ -1,4 +1,3 @@
-# COMPILE_PREFABS
 import os
 import socket
 import sys
@@ -9,14 +8,14 @@ from pathlib import Path
 from prefab_classes import prefab, attribute
 from prefab_classes.funcs import to_json
 
-from ..hotkeys import hotkey_or_none
+from .hotkeys import hotkey_or_none
 
 if getattr(sys, "frozen", False):  # pragma: nocover
     # Application is .exe, use visible files
     base_path = Path(sys.executable).parent
 else:
     # Running as .py - use standard folder structure
-    base_path = Path(__file__).parents[1]
+    base_path = Path(__file__).parent
 
 settings_file = Path(base_path / "settings.json")
 default_template_folder = Path(base_path / "templates")
@@ -33,7 +32,7 @@ except Exception:
     )
 
 
-@prefab(compile_prefab=True, compile_fallback=True)
+@prefab
 class Settings:
     """
     Global persistent settings handler
